+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [unreleased]
 
+### Changed
+
+- Use `jemalloc` on most unix and unix-like systems (using `mimalloc` on windows).
+
 ## [0.11.0] - 2025-06-14
 
 ### Added